works

+++ java/Definition.java
@@ -9,9 +9,9 @@ import java.util.*;
 public class Definition {
 	int attenuation;
 	float lineWidth;
-	double opacity;
+	float opacity;
 
-	Definition(int attenuation, float lineWidth, double opacity) {
+	Definition(int attenuation, float lineWidth, float opacity) {
 		this.attenuation = attenuation;
 		this.lineWidth = lineWidth;
 		this.opacity = opacity;
